Improve toolbar titles and suppress Debug Cell
