+++ package.nls.json
@@ -95,8 +95,8 @@
     "jupyter.command.jupyter.restartkernel.shorttitle": "Restart",
     "jupyter.command.jupyter.expandallcells.title": "Expand All Interactive Cells",
     "jupyter.command.jupyter.collapseallcells.title": "Collapse All Interactive Cells",
-    "jupyter.command.jupyter.expandallcells.shorttitle": "Expand All",
-    "jupyter.command.jupyter.collapseallcells.shorttitle": "Collapse All",
+    "jupyter.command.jupyter.expandallcells.shorttitle": "Expand",
+    "jupyter.command.jupyter.collapseallcells.shorttitle": "Collapse",
     "jupyter.command.jupyter.addcellbelow.title": "Add Empty Cell to File",
     "jupyter.command.jupyter.scrolltocell.title": "Scroll Cell Into View",
     "jupyter.command.jupyter.createnewnotebook.title": "Create New Blank Notebook",
@@ -132,6 +132,10 @@
     "DataScience.exportDialogComplete": "Notebook written to {0}",
     "DataScience.exportDialogFailed": "Failed to export notebook. {0}",
     "DataScience.exportOpenQuestion1": "Open in editor",
+    "DataScience.notebookExport": "Export",
+    "DataScience.exportAsNotebook": "Export To Jupyter Notebook",
+    "DataScience.exportAsNotebook.shorttitle": "Save",
+    "DataScience.exportAs": "Export As",
     "DataScience.notebookExportAs": "Export As",
     "DataScience.notebookExportAs.shorttitle": "Export",
     "DataScience.exportAsQuickPickPlaceholder": "Export As...",
