+++ src/client/datascience/editor-integration/codeLensFactory.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 'use strict';
 import { inject, injectable } from 'inversify';
-import { CodeLens, Command, Event, EventEmitter, Range, TextDocument, Uri } from 'vscode';
+import { CodeLens, Command, Event, EventEmitter, Range, TextDocument, Uri, workspace } from 'vscode';
 
 import { IDocumentManager, IWorkspaceService } from '../../common/application/types';
 import { traceWarning } from '../../common/logger';
@@ -353,9 +353,13 @@ export class CodeLensFactory implements ICodeLensFactory, IInteractiveWindowList
         commandName: string,
         isFirst: boolean
     ): CodeLens | undefined {
-        // Do not generate interactive window codelenses
-        // for TextDocuments which are part of NotebookDocuments
-        if (document.notebook !== undefined) {
+        // Do not generate interactive window codelenses for TextDocuments which are part of NotebookDocuments
+        // and do not generate debug codelenses for new interactive window (not properly supported yet)
+        if (
+            document.notebook !== undefined ||
+            (commandName.includes('debug') &&
+                workspace.getConfiguration('interactive.experiments').get('enable') === true)
+        ) {
             return;
         }
